Major refactoring. 1. Root level folder is learning-to-transduce. 2. Src folder is ltt. Code to be executed from root. in:root from ltt.layers import Dense
