--- layers/__init__.py
@@ -1,3 +1,0 @@
-## To expose classes, functions in the root folder, add imports here.
-from abstract_layer import AbstractLayer
-from dense import Dense
