+++ ltt/layers/__init__.py
@@ -1,0 +1,5 @@
+## To expose classes, functions in the root folder, add imports here.
+from __future__ import absolute_import
+from .abstract_layer import *
+from .dense import *
+from ..utils import array_init
