+++ ltt/layers/abstract_layer.py
