+++ ltt/utils/array_init.py
