--- utils/__init__.py
@@ -1,1 +1,0 @@
-from array_init import array_init
